refactor(DatasetExplorer): extract filter & dialog components (#121)
Extract 6 components from DatasetExplorer.tsx to reduce it from 3,513
to 2,572 lines (-941 lines). This is Phase 3 of the decomposition epic (#92).

Components extracted:
- FilterChip: individual filter pill with NOT toggle and remove
- ActiveFilters: filter bar with chips, Save/Clear buttons
- ChartSettingsMenu: chart label toggle (Counts/Percentages)
- SavedFiltersBar: preset count badge with Load/Manage buttons
- PresetDialogs: Save/Load/Manage filter preset modals
- DashboardDialogs: Save/Load/Manage dashboard modals

All components are props-based (no new context needed). Includes 11 new
tests for FilterChip and ActiveFilters. Removed 2 unused imports.
