+++ client/src/pages/DatasetExplorer/DatasetExplorer.tsx
@@ -11,7 +11,6 @@ import {
   rangeKey,
   rangesEqual,
   getFilterCountKey,
-  getFilterColumn,
   filterContainsColumn,
   migrateFiltersToCurrentSchema,
 } from '../../utils/filterHelpers'
@@ -29,9 +28,13 @@ import { TableViewChart } from './components/TableViewChart'
 import { HistogramChart } from './components/HistogramChart'
 import { SurvivalChart } from './components/SurvivalChart'
 import { MapChart } from './components/MapChart'
+import { ActiveFilters } from './components/ActiveFilters'
+import { ChartSettingsMenu } from './components/ChartSettingsMenu'
+import { SavedFiltersBar } from './components/SavedFiltersBar'
+import { PresetDialogs } from './components/PresetDialogs'
+import { DashboardDialogs } from './components/DashboardDialogs'
 import {
   MAX_PIE_CATEGORIES,
-  CHART_LABEL_STORAGE_PREFIX,
   DASHBOARD_SCOPE_KEY,
   CACHE_TTL_MS,
   CACHE_MAX_ENTRIES_PER_TABLE,
@@ -1830,1021 +1833,77 @@ function DatasetExplorer() {
             Chart settings
           </button>
           {showSettingsMenu && (
-            <div
-              ref={settingsMenuRef}
-              style={{
-                position: 'absolute',
-                top: 'calc(100% + 8px)',
-                right: 0,
-                background: 'white',
-                borderRadius: '8px',
-                boxShadow: '0 4px 12px rgba(0,0,0,0.15)',
-                padding: '0.75rem 1rem',
-                width: '220px',
-                zIndex: 5
-              }}
-            >
-              <div style={{ fontWeight: 600, fontSize: '0.85rem', marginBottom: '0.5rem' }}>Chart settings</div>
-              <div style={{ fontSize: '0.75rem', marginBottom: '0.25rem', color: '#444' }}>Chart labels</div>
-              <div style={{ display: 'flex', gap: '0.35rem' }}>
-                <button
-                  type="button"
-                  onClick={() => {
-                    setShowPercentageLabels(false)
-                    setShowSettingsMenu(false)
-                    try {
-                      localStorage.setItem(`${CHART_LABEL_STORAGE_PREFIX}${identifier}`, 'count')
-                    } catch (e) {
-                      console.error('Failed to save chart label preference', e)
-                    }
-                  }}
-                  style={{
-                    border: 'none',
-                    borderRadius: '999px',
-                    padding: '0.2rem 0.9rem',
-                    fontSize: '0.75rem',
-                    cursor: 'pointer',
-                    background: showPercentageLabels ? '#ECEFF1' : '#1976D2',
-                    color: showPercentageLabels ? '#333' : 'white'
-                  }}
-                >
-                  Counts
-                </button>
-                <button
-                  type="button"
-                  onClick={() => {
-                    setShowPercentageLabels(true)
-                    setShowSettingsMenu(false)
-                    try {
-                      localStorage.setItem(`${CHART_LABEL_STORAGE_PREFIX}${identifier}`, 'percent')
-                    } catch (e) {
-                      console.error('Failed to save chart label preference', e)
-                    }
-                  }}
-                  style={{
-                    border: 'none',
-                    borderRadius: '999px',
-                    padding: '0.2rem 0.9rem',
-                    fontSize: '0.75rem',
-                    cursor: 'pointer',
-                    background: showPercentageLabels ? '#1976D2' : '#ECEFF1',
-                    color: showPercentageLabels ? 'white' : '#333'
-                  }}
-                >
-                  Percentages
-                </button>
-              </div>
-              <div style={{ fontSize: '0.7rem', color: '#777', marginTop: '0.4rem' }}>
-                {showPercentageLabels ? 'Percentages may exceed 100% when parents overlap.' : 'Switch to percentages when needed.'}
-              </div>
-            </div>
+            <ChartSettingsMenu
+              showPercentageLabels={showPercentageLabels}
+              setShowPercentageLabels={setShowPercentageLabels}
+              onClose={() => setShowSettingsMenu(false)}
+              identifier={identifier}
+              menuRef={settingsMenuRef}
+            />
           )}
         </div>
       </div>
 
-      {/* Saved Filters Bar - Always visible when presets exist */}
-      {presets.length > 0 && (
-        <div style={{
-          marginBottom: '1rem',
-          background: '#E3F2FD',
-          padding: '0.75rem 1rem',
-          borderRadius: '8px',
-          border: '1px solid #90CAF9',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'space-between'
-        }}>
-          <div style={{ fontSize: '0.875rem', color: '#1976D2', fontWeight: 500 }}>
-            {presets.length} saved filter{presets.length !== 1 ? 's' : ''}
-          </div>
-          <div style={{ display: 'flex', gap: '0.5rem' }}>
-            <div style={{ position: 'relative' }}>
-              <button
-                onClick={() => setShowPresetsDropdown(!showPresetsDropdown)}
-                style={{
-                  padding: '0.25rem 0.75rem',
-                  background: '#2196F3',
-                  color: 'white',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  fontSize: '0.75rem'
-                }}
-                title="Load a saved filter"
-              >
-                Load Filter
-              </button>
-            </div>
-            <button
-              onClick={() => setShowManagePresetsDialog(true)}
-              style={{
-                padding: '0.25rem 0.75rem',
-                background: '#FF9800',
-                color: 'white',
-                border: 'none',
-                borderRadius: '4px',
-                cursor: 'pointer',
-                fontSize: '0.75rem'
-              }}
-              title="Manage saved filters"
-            >
-              Manage
-            </button>
-          </div>
-        </div>
-      )}
+      <SavedFiltersBar
+        presets={presets}
+        showPresetsDropdown={showPresetsDropdown}
+        setShowPresetsDropdown={setShowPresetsDropdown}
+        setShowManagePresetsDialog={setShowManagePresetsDialog}
+      />
 
-      {/* Active Filters */}
-      {filters.length > 0 && (
-        <div style={{
-          marginBottom: '1rem',
-          background: '#F5F5F5',
-          padding: '1rem',
-          borderRadius: '8px',
-          border: '1px solid #E0E0E0'
-        }}>
-          <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', marginBottom: '0.5rem' }}>
-            <strong style={{ fontSize: '0.875rem' }}>Active Filters:</strong>
-            <div style={{ display: 'flex', gap: '0.5rem', alignItems: 'center' }}>
-              <button
-                onClick={() => setShowSavePresetDialog(true)}
-                style={{
-                  padding: '0.25rem 0.75rem',
-                  background: '#4CAF50',
-                  color: 'white',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  fontSize: '0.75rem'
-                }}
-                title="Save current filters"
-              >
-                Save Filter
-              </button>
-              <button
-                onClick={clearFilters}
-                style={{
-                  padding: '0.25rem 0.75rem',
-                  background: '#f44336',
-                  color: 'white',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  fontSize: '0.75rem'
-                }}
-              >
-                Clear All
-              </button>
-            </div>
-          </div>
-          <div style={{ display: 'flex', gap: '0.5rem', flexWrap: 'wrap', alignItems: 'center' }}>
-            {filters.map((filter, idx) => {
-              const actualFilter = unwrapNot(filter)
-              if (!actualFilter) return null
-              const isNot = !!filter.not
+      <ActiveFilters
+        filters={filters}
+        setFilters={setFilters}
+        onSaveFilter={() => setShowSavePresetDialog(true)}
+        onClearFilters={clearFilters}
+        getTableColor={getTableColor}
+        getFilterTableNameForCacheKey={getFilterTableNameForCacheKey}
+        targetFromCacheKey={targetFromCacheKey}
+        formatRangeValue={formatRangeValue}
+        clearColumnFilter={clearColumnFilter}
+        datasetTables={dataset?.tables}
+      />
 
-              const columnName = getFilterColumn(actualFilter)
-              const tableName = getFilterTableNameForCacheKey(actualFilter)
-              const tableColor = tableName ? getTableColor(tableName) : '#9E9E9E'
-              const table = dataset?.tables.find(t => t.name === tableName)
+      <PresetDialogs
+        presets={presets}
+        showSavePresetDialog={showSavePresetDialog}
+        setShowSavePresetDialog={setShowSavePresetDialog}
+        showPresetsDropdown={showPresetsDropdown}
+        setShowPresetsDropdown={setShowPresetsDropdown}
+        showManagePresetsDialog={showManagePresetsDialog}
+        setShowManagePresetsDialog={setShowManagePresetsDialog}
+        presetNameInput={presetNameInput}
+        setPresetNameInput={setPresetNameInput}
+        editingPresetId={editingPresetId}
+        setEditingPresetId={setEditingPresetId}
+        savePreset={savePreset}
+        applyPreset={applyPreset}
+        deletePreset={deletePreset}
+        renamePreset={renamePreset}
+        exportPresets={exportPresets}
+        importPresets={importPresets}
+      />
 
-              // Extract count-by table for border color
-              const countByTable = targetFromCacheKey(filter.countByKey)
-              const countByColor = countByTable ? getTableColor(countByTable) : tableColor
 
-              let displayValue = String(actualFilter.value)
-              let logicType = '' // For tooltip
-
-              // Remove handler - uses actualFilter's column/table regardless of NOT wrapper
-              const removeHandler = () => {
-                if (tableName && columnName) {
-                  clearColumnFilter(tableName, columnName, getFilterCountKey(filter))
-                } else {
-                  setFilters(filters.filter((_, i) => i !== idx))
-                }
-              }
-
-              // Toggle NOT wrapper
-              const toggleNot = () => {
-                setFilters(prev => {
-                  const updated = [...prev]
-                  if (isNot) {
-                    // Remove NOT wrapper
-                    updated[idx] = actualFilter
-                  } else {
-                    // Add NOT wrapper
-                    updated[idx] = { not: actualFilter }
-                  }
-                  return updated
-                })
-              }
-
-              if (actualFilter.operator === 'between' && Array.isArray(actualFilter.value)) {
-                displayValue = `[${typeof actualFilter.value[0] === 'number' ? actualFilter.value[0].toFixed(2) : actualFilter.value[0]}, ${typeof actualFilter.value[1] === 'number' ? actualFilter.value[1].toFixed(2) : actualFilter.value[1]}]`
-                logicType = 'Range'
-              } else if (actualFilter.operator === 'in' && Array.isArray(actualFilter.value)) {
-                const displayVals = actualFilter.value.map(v => {
-                  if (v === '') return '(Empty)'
-                  if (v === ' ') return '(Space)'
-                  return v
-                })
-                // Show OR for multi-value selections
-                if (actualFilter.value.length > 1) {
-                  displayValue = displayVals.slice(0, 3).join(' OR ')
-                  if (actualFilter.value.length > 3) {
-                    displayValue += ` OR ${actualFilter.value.length - 3} more...`
-                  }
-                } else {
-                  displayValue = displayVals[0] || ''
-                }
-                logicType = actualFilter.value.length > 1 ? `OR (${actualFilter.value.length} values)` : 'Single value'
-              } else if (actualFilter.operator === 'eq') {
-                if (actualFilter.value === '') displayValue = '(Empty)'
-                else if (actualFilter.value === ' ') displayValue = '(Space)'
-                else displayValue = String(actualFilter.value)
-                logicType = 'Equals'
-              } else if (actualFilter.or && Array.isArray(actualFilter.or)) {
-                const ranges = actualFilter.or
-                  .map(rangeFilter => rangeFilter as Filter)
-                  .filter(rangeFilter => rangeFilter.column === actualFilter.column && rangeFilter.operator === 'between' && Array.isArray(rangeFilter.value))
-                  .map(rangeFilter => {
-                    const [start, end] = rangeFilter.value
-                    const startLabel = typeof start === 'number' ? formatRangeValue(start) : String(start)
-                    const endLabel = typeof end === 'number' ? formatRangeValue(end) : String(end)
-                    return `${startLabel}–${endLabel}`
-                  })
-
-                displayValue = ranges.join(' OR ')
-                logicType = `OR (${ranges.length} ranges)`
-              }
-              const columnLabel = columnName ?? '(Column)'
-              const notPrefix = isNot ? 'NOT: ' : ''
-              const tooltipText = tableName
-                ? `${table?.displayName || tableName}.${columnLabel}\n${notPrefix}${logicType}\nValue: ${displayValue}`
-                : columnLabel
-
-              const showAndSeparator = idx > 0
-
-              return (
-                <React.Fragment key={idx}>
-                  {showAndSeparator && (
-                    <div style={{
-                      color: '#666',
-                      fontSize: '0.75rem',
-                      fontWeight: 600,
-                      padding: '0 0.25rem',
-                      userSelect: 'none'
-                    }}>
-                      AND
-                    </div>
-                  )}
-                  <div
-                    style={{
-                      background: isNot ? `linear-gradient(135deg, ${tableColor}DD, ${tableColor}BB)` : tableColor,
-                      padding: '0.25rem 0.75rem',
-                      borderRadius: '4px',
-                      fontSize: '0.875rem',
-                      display: 'flex',
-                      alignItems: 'center',
-                      gap: '0.5rem',
-                      outline: `4px solid ${countByColor}`,
-                      outlineOffset: '2px',
-                      border: isNot ? `2px dashed rgba(255,255,255,0.6)` : 'none',
-                      color: 'white',
-                      fontWeight: 500,
-                      opacity: isNot ? 0.9 : 1
-                    }}
-                    title={tooltipText}
-                  >
-                    {isNot && (
-                      <span style={{
-                        background: 'rgba(0,0,0,0.3)',
-                        padding: '0.1rem 0.35rem',
-                        borderRadius: '3px',
-                        fontSize: '0.7rem',
-                        fontWeight: 700,
-                        marginRight: '0.1rem'
-                      }}>
-                        NOT
-                      </span>
-                    )}
-                    <span style={{ textDecoration: isNot ? 'line-through' : 'none' }}>
-                      <strong>{columnLabel}:</strong> {displayValue}
-                    </span>
-                    <button
-                      onClick={toggleNot}
-                      style={{
-                        background: isNot ? 'rgba(0,0,0,0.3)' : 'rgba(255,255,255,0.3)',
-                        border: 'none',
-                        color: 'white',
-                        cursor: 'pointer',
-                        padding: '0 0.3rem',
-                        fontSize: '0.75rem',
-                        lineHeight: '1',
-                        borderRadius: '3px',
-                        fontWeight: 'bold'
-                      }}
-                      title={isNot ? 'Remove NOT' : 'Add NOT'}
-                    >
-                      ¬
-                    </button>
-                    <button
-                      onClick={removeHandler}
-                      style={{
-                        background: 'rgba(255,255,255,0.3)',
-                        border: 'none',
-                        color: 'white',
-                        cursor: 'pointer',
-                        padding: '0 0.25rem',
-                        fontSize: '1rem',
-                        lineHeight: '1',
-                        borderRadius: '3px',
-                        fontWeight: 'bold'
-                      }}
-                    >
-                      ×
-                    </button>
-                  </div>
-                </React.Fragment>
-              )
-            })}
-          </div>
-        </div>
-      )}
-
-      {/* Save Filter Dialog */}
-      {showSavePresetDialog && (
-        <div
-          style={{
-            position: 'fixed',
-            top: 0,
-            left: 0,
-            right: 0,
-            bottom: 0,
-            background: 'rgba(0,0,0,0.5)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            zIndex: 1000
-          }}
-          onClick={() => setShowSavePresetDialog(false)}
-        >
-          <div
-            style={{
-              background: 'white',
-              padding: '1.5rem',
-              borderRadius: '8px',
-              minWidth: '400px',
-              boxShadow: '0 4px 12px rgba(0,0,0,0.2)'
-            }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            <h3 style={{ marginTop: 0 }}>Save Filter</h3>
-            <input
-              type="text"
-              value={presetNameInput}
-              onChange={(e) => setPresetNameInput(e.target.value)}
-              placeholder="Enter filter name..."
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') savePreset()
-                if (e.key === 'Escape') setShowSavePresetDialog(false)
-              }}
-              autoFocus
-              style={{
-                width: '100%',
-                padding: '0.5rem',
-                marginBottom: '1rem',
-                border: '1px solid #ddd',
-                borderRadius: '4px',
-                fontSize: '0.875rem'
-              }}
-            />
-            <div style={{ display: 'flex', gap: '0.5rem', justifyContent: 'flex-end' }}>
-              <button
-                onClick={() => setShowSavePresetDialog(false)}
-                style={{
-                  padding: '0.5rem 1rem',
-                  background: '#f0f0f0',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  fontSize: '0.875rem'
-                }}
-              >
-                Cancel
-              </button>
-              <button
-                onClick={savePreset}
-                disabled={!presetNameInput.trim()}
-                style={{
-                  padding: '0.5rem 1rem',
-                  background: presetNameInput.trim() ? '#4CAF50' : '#ccc',
-                  color: 'white',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: presetNameInput.trim() ? 'pointer' : 'not-allowed',
-                  fontSize: '0.875rem'
-                }}
-              >
-                Save
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Load Filter Dropdown */}
-      {showPresetsDropdown && (
-        <div
-          style={{
-            position: 'fixed',
-            top: 0,
-            left: 0,
-            right: 0,
-            bottom: 0,
-            zIndex: 999
-          }}
-          onClick={() => setShowPresetsDropdown(false)}
-        >
-          <div
-            style={{
-              position: 'absolute',
-              top: '120px',
-              right: '20px',
-              background: 'white',
-              border: '1px solid #ddd',
-              borderRadius: '8px',
-              boxShadow: '0 4px 12px rgba(0,0,0,0.15)',
-              padding: '0.5rem',
-              minWidth: '300px',
-              maxHeight: '400px',
-              overflowY: 'auto'
-            }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            <div style={{ padding: '0.5rem', borderBottom: '1px solid #eee', marginBottom: '0.5rem' }}>
-              <strong style={{ fontSize: '0.875rem' }}>Select Filter</strong>
-            </div>
-            {presets.map((preset) => (
-              <div
-                key={preset.id}
-                onClick={() => applyPreset(preset)}
-                style={{
-                  padding: '0.75rem',
-                  cursor: 'pointer',
-                  borderRadius: '4px',
-                  marginBottom: '0.25rem',
-                  border: '1px solid #eee',
-                  transition: 'background 0.2s'
-                }}
-                onMouseEnter={(e) => e.currentTarget.style.background = '#f0f0f0'}
-                onMouseLeave={(e) => e.currentTarget.style.background = 'white'}
-              >
-                <div style={{ fontWeight: 500, fontSize: '0.875rem', marginBottom: '0.25rem' }}>
-                  {preset.name}
-                </div>
-                <div style={{ fontSize: '0.7rem', color: '#666' }}>
-                  {preset.filters.length} filter{preset.filters.length !== 1 ? 's' : ''} · {new Date(preset.createdAt).toLocaleDateString()}
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-
-      {/* Manage Filters Dialog */}
-      {showManagePresetsDialog && (
-        <div
-          style={{
-            position: 'fixed',
-            top: 0,
-            left: 0,
-            right: 0,
-            bottom: 0,
-            background: 'rgba(0,0,0,0.5)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            zIndex: 1000
-          }}
-          onClick={() => setShowManagePresetsDialog(false)}
-        >
-          <div
-            style={{
-              background: 'white',
-              padding: '1.5rem',
-              borderRadius: '8px',
-              minWidth: '500px',
-              maxHeight: '600px',
-              overflowY: 'auto',
-              boxShadow: '0 4px 12px rgba(0,0,0,0.2)'
-            }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: '1rem' }}>
-              <h3 style={{ margin: 0 }}>Manage Saved Filters</h3>
-              <div style={{ display: 'flex', gap: '0.5rem' }}>
-                <button
-                  onClick={exportPresets}
-                  style={{
-                    padding: '0.4rem 0.75rem',
-                    background: '#2196F3',
-                    color: 'white',
-                    border: 'none',
-                    borderRadius: '4px',
-                    cursor: 'pointer',
-                    fontSize: '0.75rem'
-                  }}
-                >
-                  Export
-                </button>
-                <label style={{
-                  padding: '0.4rem 0.75rem',
-                  background: '#FF9800',
-                  color: 'white',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  fontSize: '0.75rem'
-                }}>
-                  Import
-                  <input
-                    type="file"
-                    accept=".json"
-                    onChange={importPresets}
-                    style={{ display: 'none' }}
-                  />
-                </label>
-              </div>
-            </div>
-            {presets.length === 0 ? (
-              <p style={{ color: '#999', textAlign: 'center', padding: '2rem' }}>
-                No filters saved yet.
-              </p>
-            ) : (
-              <div style={{ display: 'flex', flexDirection: 'column', gap: '0.5rem' }}>
-                {presets.map((preset) => (
-                  <div
-                    key={preset.id}
-                    style={{
-                      border: '1px solid #ddd',
-                      borderRadius: '6px',
-                      padding: '0.75rem'
-                    }}
-                  >
-                    {editingPresetId === preset.id ? (
-                      <div style={{ display: 'flex', gap: '0.5rem', marginBottom: '0.5rem' }}>
-                        <input
-                          type="text"
-                          defaultValue={preset.name}
-                          onKeyDown={(e) => {
-                            if (e.key === 'Enter') renamePreset(preset.id, e.currentTarget.value)
-                            if (e.key === 'Escape') setEditingPresetId(null)
-                          }}
-                          autoFocus
-                          style={{
-                            flex: 1,
-                            padding: '0.25rem 0.5rem',
-                            border: '1px solid #2196F3',
-                            borderRadius: '4px',
-                            fontSize: '0.875rem'
-                          }}
-                        />
-                        <button
-                          onClick={(e) => renamePreset(preset.id, (e.currentTarget.previousElementSibling as HTMLInputElement | null)?.value || '')}
-                          style={{
-                            padding: '0.25rem 0.5rem',
-                            background: '#4CAF50',
-                            color: 'white',
-                            border: 'none',
-                            borderRadius: '4px',
-                            cursor: 'pointer',
-                            fontSize: '0.75rem'
-                          }}
-                        >
-                          Save
-                        </button>
-                        <button
-                          onClick={() => setEditingPresetId(null)}
-                          style={{
-                            padding: '0.25rem 0.5rem',
-                            background: '#f0f0f0',
-                            border: 'none',
-                            borderRadius: '4px',
-                            cursor: 'pointer',
-                            fontSize: '0.75rem'
-                          }}
-                        >
-                          Cancel
-                        </button>
-                      </div>
-                    ) : (
-                      <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: '0.5rem' }}>
-                        <strong style={{ fontSize: '0.875rem' }}>{preset.name}</strong>
-                        <div style={{ display: 'flex', gap: '0.5rem' }}>
-                          <button
-                            onClick={() => setEditingPresetId(preset.id)}
-                            style={{
-                              padding: '0.25rem 0.5rem',
-                              background: '#2196F3',
-                              color: 'white',
-                              border: 'none',
-                              borderRadius: '4px',
-                              cursor: 'pointer',
-                              fontSize: '0.75rem'
-                            }}
-                          >
-                            Rename
-                          </button>
-                          <button
-                            onClick={() => {
-                              if (window.confirm(`Delete saved filter "${preset.name}"?`)) {
-                                deletePreset(preset.id)
-                              }
-                            }}
-                            style={{
-                              padding: '0.25rem 0.5rem',
-                              background: '#f44336',
-                              color: 'white',
-                              border: 'none',
-                              borderRadius: '4px',
-                              cursor: 'pointer',
-                              fontSize: '0.75rem'
-                            }}
-                          >
-                            Delete
-                          </button>
-                        </div>
-                      </div>
-                    )}
-                    <div style={{ fontSize: '0.75rem', color: '#666' }}>
-                      {preset.filters.length} filter{preset.filters.length !== 1 ? 's' : ''} · Created {new Date(preset.createdAt).toLocaleDateString()}
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-            <div style={{ marginTop: '1rem', display: 'flex', justifyContent: 'flex-end' }}>
-              <button
-                onClick={() => setShowManagePresetsDialog(false)}
-                style={{
-                  padding: '0.5rem 1rem',
-                  background: '#f0f0f0',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  fontSize: '0.875rem'
-                }}
-              >
-                Close
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Save Dashboard Dialog */}
-      {showSaveDashboardDialog && (
-        <div
-          style={{
-            position: 'fixed',
-            top: 0,
-            left: 0,
-            right: 0,
-            bottom: 0,
-            background: 'rgba(0,0,0,0.5)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            zIndex: 1000
-          }}
-          onClick={() => setShowSaveDashboardDialog(false)}
-        >
-          <div
-            style={{
-              background: 'white',
-              padding: '1.5rem',
-              borderRadius: '8px',
-              minWidth: '400px',
-              boxShadow: '0 4px 12px rgba(0,0,0,0.2)'
-            }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            <h3 style={{ marginTop: 0 }}>Save Dashboard</h3>
-            <input
-              type="text"
-              value={newDashboardName}
-              onChange={(e) => setNewDashboardName(e.target.value)}
-              placeholder="Enter dashboard name..."
-              onKeyDown={(e) => {
-                if (e.key === 'Enter' && newDashboardName.trim()) saveDashboard(newDashboardName.trim())
-                if (e.key === 'Escape') setShowSaveDashboardDialog(false)
-              }}
-              autoFocus
-              style={{
-                width: '100%',
-                padding: '0.5rem',
-                marginBottom: '1rem',
-                border: '1px solid #ddd',
-                borderRadius: '4px',
-                fontSize: '0.875rem'
-              }}
-            />
-            <div style={{ display: 'flex', gap: '0.5rem', justifyContent: 'flex-end' }}>
-              <button
-                onClick={() => {
-                  setShowSaveDashboardDialog(false)
-                  setNewDashboardName('')
-                }}
-                style={{
-                  padding: '0.5rem 1rem',
-                  background: '#f0f0f0',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  fontSize: '0.875rem'
-                }}
-              >
-                Cancel
-              </button>
-              <button
-                onClick={() => saveDashboard(newDashboardName.trim())}
-                disabled={!newDashboardName.trim()}
-                style={{
-                  padding: '0.5rem 1rem',
-                  background: newDashboardName.trim() ? '#4CAF50' : '#ccc',
-                  color: 'white',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: newDashboardName.trim() ? 'pointer' : 'not-allowed',
-                  fontSize: '0.875rem'
-                }}
-              >
-                Save
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Load Dashboard Dialog */}
-      {showLoadDashboardDialog && (
-        <div
-          style={{
-            position: 'fixed',
-            top: 0,
-            left: 0,
-            right: 0,
-            bottom: 0,
-            background: 'rgba(0,0,0,0.5)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            zIndex: 1000
-          }}
-          onClick={() => setShowLoadDashboardDialog(false)}
-        >
-          <div
-            style={{
-              background: 'white',
-              padding: '1.5rem',
-              borderRadius: '8px',
-              minWidth: '400px',
-              maxWidth: '500px',
-              maxHeight: '600px',
-              overflowY: 'auto',
-              boxShadow: '0 4px 12px rgba(0,0,0,0.2)'
-            }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            <h3 style={{ marginTop: 0 }}>Load Dashboard</h3>
-            {savedDashboards.length === 0 ? (
-              <p style={{ color: '#999', textAlign: 'center', padding: '2rem' }}>
-                No saved dashboards yet.
-              </p>
-            ) : (
-              <div style={{ display: 'flex', flexDirection: 'column', gap: '0.5rem' }}>
-                {savedDashboards.map(dashboard => (
-                  <div
-                    key={dashboard.id}
-                    onClick={() => {
-                      loadDashboard(dashboard.id)
-                      setShowLoadDashboardDialog(false)
-                    }}
-                    style={{
-                      padding: '0.75rem',
-                      border: activeDashboardId === dashboard.id ? '2px solid #2196F3' : '1px solid #ddd',
-                      borderRadius: '6px',
-                      cursor: 'pointer',
-                      transition: 'background 0.2s, border-color 0.2s',
-                      background: activeDashboardId === dashboard.id ? '#E3F2FD' : 'white'
-                    }}
-                    onMouseEnter={(e) => {
-                      if (activeDashboardId !== dashboard.id) {
-                        e.currentTarget.style.background = '#f5f5f5'
-                        e.currentTarget.style.borderColor = '#2196F3'
-                      }
-                    }}
-                    onMouseLeave={(e) => {
-                      if (activeDashboardId !== dashboard.id) {
-                        e.currentTarget.style.background = 'white'
-                        e.currentTarget.style.borderColor = '#ddd'
-                      }
-                    }}
-                  >
-                    <div style={{ fontWeight: 500, fontSize: '0.875rem', marginBottom: '0.25rem' }}>
-                      {dashboard.name}
-                      {activeDashboardId === dashboard.id && (
-                        <span style={{ marginLeft: '0.5rem', color: '#2196F3', fontSize: '0.75rem' }}>(Most Recently Loaded)</span>
-                      )}
-                    </div>
-                    <div style={{ fontSize: '0.75rem', color: '#666' }}>
-                      {dashboard.charts.length} chart{dashboard.charts.length !== 1 ? 's' : ''} · Created {new Date(dashboard.createdAt).toLocaleDateString()}
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-            <div style={{ marginTop: '1rem', display: 'flex', justifyContent: 'flex-end' }}>
-              <button
-                onClick={() => setShowLoadDashboardDialog(false)}
-                style={{
-                  padding: '0.5rem 1rem',
-                  background: '#f0f0f0',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  fontSize: '0.875rem'
-                }}
-              >
-                Close
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Manage Dashboards Dialog */}
-      {showManageDashboardsDialog && (
-        <div
-          style={{
-            position: 'fixed',
-            top: 0,
-            left: 0,
-            right: 0,
-            bottom: 0,
-            background: 'rgba(0,0,0,0.5)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            zIndex: 1000
-          }}
-          onClick={() => setShowManageDashboardsDialog(false)}
-        >
-          <div
-            style={{
-              background: 'white',
-              padding: '1.5rem',
-              borderRadius: '8px',
-              minWidth: '500px',
-              maxHeight: '600px',
-              overflowY: 'auto',
-              boxShadow: '0 4px 12px rgba(0,0,0,0.2)'
-            }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            <h3 style={{ margin: '0 0 1rem 0' }}>Manage Saved Dashboards</h3>
-            {savedDashboards.length === 0 ? (
-              <p style={{ color: '#999', textAlign: 'center', padding: '2rem' }}>
-                No dashboards saved yet.
-              </p>
-            ) : (
-              <div style={{ display: 'flex', flexDirection: 'column', gap: '0.5rem' }}>
-                {savedDashboards.map((dashboard) => (
-                  <div
-                    key={dashboard.id}
-                    style={{
-                      border: '1px solid #ddd',
-                      borderRadius: '6px',
-                      padding: '0.75rem'
-                    }}
-                  >
-                    {editingDashboardId === dashboard.id ? (
-                      <div style={{ display: 'flex', gap: '0.5rem', marginBottom: '0.5rem' }}>
-                        <input
-                          type="text"
-                          defaultValue={dashboard.name}
-                          onKeyDown={(e) => {
-                            if (e.key === 'Enter') renameDashboard(dashboard.id, e.currentTarget.value)
-                            if (e.key === 'Escape') setEditingDashboardId(null)
-                          }}
-                          autoFocus
-                          style={{
-                            flex: 1,
-                            padding: '0.25rem 0.5rem',
-                            border: '1px solid #2196F3',
-                            borderRadius: '4px',
-                            fontSize: '0.875rem'
-                          }}
-                        />
-                        <button
-                          onClick={(e) => {
-                            const input = e.currentTarget.previousElementSibling as HTMLInputElement
-                            renameDashboard(dashboard.id, input?.value || '')
-                          }}
-                          style={{
-                            padding: '0.25rem 0.5rem',
-                            background: '#4CAF50',
-                            color: 'white',
-                            border: 'none',
-                            borderRadius: '4px',
-                            cursor: 'pointer',
-                            fontSize: '0.75rem'
-                          }}
-                        >
-                          Save
-                        </button>
-                        <button
-                          onClick={() => setEditingDashboardId(null)}
-                          style={{
-                            padding: '0.25rem 0.5rem',
-                            background: '#f0f0f0',
-                            border: 'none',
-                            borderRadius: '4px',
-                            cursor: 'pointer',
-                            fontSize: '0.75rem'
-                          }}
-                        >
-                          Cancel
-                        </button>
-                      </div>
-                    ) : (
-                      <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: '0.5rem' }}>
-                        <strong style={{ fontSize: '0.875rem' }}>{dashboard.name}</strong>
-                        <div style={{ display: 'flex', gap: '0.5rem' }}>
-                          <button
-                            onClick={() => {
-                              setEditingDashboardId(dashboard.id)
-                              setEditingDashboardName(dashboard.name)
-                            }}
-                            style={{
-                              padding: '0.25rem 0.5rem',
-                              background: '#2196F3',
-                              color: 'white',
-                              border: 'none',
-                              borderRadius: '4px',
-                              cursor: 'pointer',
-                              fontSize: '0.75rem'
-                            }}
-                          >
-                            Rename
-                          </button>
-                          <button
-                            onClick={() => {
-                              if (window.confirm(`Delete dashboard "${dashboard.name}"?`)) {
-                                deleteDashboard(dashboard.id)
-                              }
-                            }}
-                            style={{
-                              padding: '0.25rem 0.5rem',
-                              background: '#f44336',
-                              color: 'white',
-                              border: 'none',
-                              borderRadius: '4px',
-                              cursor: 'pointer',
-                              fontSize: '0.75rem'
-                            }}
-                          >
-                            Delete
-                          </button>
-                        </div>
-                      </div>
-                    )}
-                    <div style={{ fontSize: '0.75rem', color: '#666' }}>
-                      {dashboard.charts.length} chart{dashboard.charts.length !== 1 ? 's' : ''} · Created {new Date(dashboard.createdAt).toLocaleDateString()}
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-            <div style={{ marginTop: '1rem', display: 'flex', justifyContent: 'flex-end' }}>
-              <button
-                onClick={() => setShowManageDashboardsDialog(false)}
-                style={{
-                  padding: '0.5rem 1rem',
-                  background: '#f0f0f0',
-                  border: 'none',
-                  borderRadius: '4px',
-                  cursor: 'pointer',
-                  fontSize: '0.875rem'
-                }}
-              >
-                Close
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
+      <DashboardDialogs
+        savedDashboards={savedDashboards}
+        activeDashboardId={activeDashboardId}
+        showSaveDashboardDialog={showSaveDashboardDialog}
+        setShowSaveDashboardDialog={setShowSaveDashboardDialog}
+        showLoadDashboardDialog={showLoadDashboardDialog}
+        setShowLoadDashboardDialog={setShowLoadDashboardDialog}
+        showManageDashboardsDialog={showManageDashboardsDialog}
+        setShowManageDashboardsDialog={setShowManageDashboardsDialog}
+        newDashboardName={newDashboardName}
+        setNewDashboardName={setNewDashboardName}
+        editingDashboardId={editingDashboardId}
+        setEditingDashboardId={setEditingDashboardId}
+        setEditingDashboardName={setEditingDashboardName}
+        saveDashboard={saveDashboard}
+        loadDashboard={loadDashboard}
+        deleteDashboard={deleteDashboard}
+        renameDashboard={renameDashboard}
+      />
 
       {/* Tab Navigation */}
       <div style={{
